Add files via upload
Project Output

The final output of the project is an interactive web interface designed for real-time fraud detection. Users can input transaction details such as transaction type, amount, sender and receiver balances before and after the transaction.

Once the “Predict” button is clicked, the trained machine learning model analyzes the provided inputs and generates a prediction result:

Prediction: '1' – Indicates that the transaction is likely to be fraudulent.

Prediction: '0' – Indicates that the transaction is genuine.

In the displayed result, the application shows a clear and user-friendly message:

“This transaction can be fraud”

The message is visually highlighted in red to alert the user to potential fraudulent activity.

This web interface provides an intuitive way to test and visualize the model’s predictions, demonstrating how machine learning can be effectively used in real-world financial transaction monitoring systems.